Add resmon
This adds a new tool, resmon, to monitor resource consumption in Spectrum
machines. The tool works by hooking up to the devlink_hwmsg tracepoint and
inspecting all EMADs coming from the Spectrum FW. The interesting ones it
further dissects to figure out what resource they describe, and then keeps
a tally count of how much of which type of resource was allocated.

The user-space tool can install and uninstall the resident BPF program.
While the BPF program is running, the user-space tool can inspect the
reported counters and dump maps that the BPF tool uses to keep track of the
individual resources.

Currently only RALUE is supported. More resources and more functionality
will be added later.

Signed-off-by: Petr Machata <petrm@nvidia.com>