resmon: Support TCAM regions [xxx]
TCAM regions are allocated through the PTAR register. PTAR.tcam_region_info
in the response EMAD holds the unique key of the allocated or freed
resource. Number of KVD slots to consume is determined from
PTAR.flexible_keys in allocation requests. In deletion requests, no keys
are passed, and thus we need to rely solely on the KVDA in the PTAR map.

[xxx] map size

Signed-off-by: Petr Machata <petrm@nvidia.com>