resmon: Add a base for a resource-monitor tool
Lay groundwork for a new demon that will count resources consumed in the
Spectrum ASIC.

The overall architecture is that of a daemon that installs a BPF program,
which monitors EMADs that the Spectrum firmware sent to the driver. Of
those it picks out a subset that might indicate that a resource has been
consumed, and pushes these to a demon for more detailed processing. The
daemon then parses these EMADs, and does the necessary bookkeeping to
know how many resources are consumed.

The individual modules that this patch introduces are as follows:

- resmon: The command-line wrapper.

- resmon-d, resmon-c: Respectively, the daemon and the client. The daemon
  opens a Unix domain socket and awaits instructions formatted as JSON-RPC
  requests. The client sends an appropriate request depending on the
  command-line command. In this patch, only the CLI "stop" and the
  corresponding JSON-RPC request are implemented.

- resmon-sock, resmon-jrpc: These modules contain helpers for handling,
  respectively, Unix sockets and JSON-RPC.

- resmon-back: The backend. This module abstracts away specifics of
  EMAD collection. In this patch, a "mock" backend is introduced, which
  will allow fully isolated testing of resmon (but currently does not
  do anything).

Signed-off-by: Petr Machata <petrm@nvidia.com>