resmon: Add test for the tool
Test resmon behavior using method "emad" which allows to send EMADs to
the daemon.

Send EMADs and verify that the appropriate counters are updated.

The test can be started using "make test" in the resmon subdirectory, which
verifies that resmon was already compiled, and if not, compiles it before.

"make resmon/test" in the top-level directory runs the test as well by way
of the usual dispatch rules. A new top-level target, "test", has been added
as well, which now dispatches to resmon/test (but is meant to collect more
targets as those become available).

Signed-off-by: Amit Cohen <amcohen@nvidia.com>