resmon: Introduce HW backend
The HW backend will expect that resmon is running on a Spectrum-type
machine and will install BPF to capture interesting EMADs, and probe
devlink to discover resource pool size.

In this patch, introduce a new module, devlink-dl, which handles the
devlink interface. Further introduce the HW backend, and implement the
get_capacity callback that dispatches to the devlink-dl module.

The new backend is not accessible at this point yet.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Co-developed-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>